LayoutTests/http/wpt/web-animations/timing-model/animations/set-the-timeline-of-an-animation-expected.txt


FAIL After setting timeline on paused animation it is still paused assert_equals: expected (string) "paused" but got (undefined) undefined
FAIL After setting timeline on animation paused outside active interval it is still paused assert_equals: expected (string) "paused" but got (undefined) undefined
FAIL After setting timeline on an idle animation without a start time it is still idle assert_equals: expected (string) "idle" but got (undefined) undefined
FAIL After setting timeline on an idle animation with a start time it is running assert_equals: expected (string) "idle" but got (undefined) undefined
FAIL After setting timeline on an idle animation with a sufficiently ancient start time it is finished assert_equals: expected (string) "idle" but got (undefined) undefined
FAIL After setting timeline on a play-pending animation it is still pending animation.play is not a function. (In 'animation.play()', 'animation.play' is undefined)
FAIL After setting timeline on a play-pending animation it begins playing after pending animation.play is not a function. (In 'animation.play()', 'animation.play' is undefined)
FAIL After setting timeline on a pause-pending animation it is still pending animation.pause is not a function. (In 'animation.pause()', 'animation.pause' is undefined)
FAIL After setting timeline on a pause-pending animation it becomes paused after pending animation.pause is not a function. (In 'animation.pause()', 'animation.pause' is undefined)
FAIL After clearing timeline on paused animation it is still paused assert_equals: expected (string) "paused" but got (undefined) undefined
FAIL After clearing timeline on finished animation it is idle assert_equals: expected (string) "finished" but got (undefined) undefined
FAIL After clearing timeline on running animation it is idle assert_equals: expected (string) "running" but got (undefined) undefined
FAIL After clearing timeline on idle animation it is still idle assert_equals: expected (string) "idle" but got (undefined) undefined
FAIL After clearing timeline on play-pending animation it is still pending createDiv(t).animate is not a function. (In 'createDiv(t).animate(null, 100 * MS_PER_SEC)', 'createDiv(t).animate' is undefined)
FAIL After clearing and re-setting timeline on play-pending animation it begins to play createDiv(t).animate is not a function. (In 'createDiv(t).animate(null, 100 * MS_PER_SEC)', 'createDiv(t).animate' is undefined)
FAIL After clearing timeline on a pause-pending animation it is still pending animation.pause is not a function. (In 'animation.pause()', 'animation.pause' is undefined)
FAIL After clearing and re-setting timeline on a pause-pending animation it becomes paused animation.pause is not a function. (In 'animation.pause()', 'animation.pause' is undefined)
FAIL After clearing and re-setting timeline on an animation in the middle of an aborted pause, it continues playing using the same start time animation.pause is not a function. (In 'animation.pause()', 'animation.pause' is undefined)
